ApiInstagram/requirements.txt
boto3==1.21.46
botocore==1.24.46
numpy==1.17.4
pandas==0.25.3
python_dateutil==2.8.2
pytz==2019.3
requests==2.22.0
typing_extensions==4.2.0
